svr: txs: move merkle methods
